// src/enums/preStressKeyType.ts
export enum PreStressKeyType {
  Id = 'preStressId',
  PreStressForce = 'preStressPreStressForce',
  PreStressSteelClass = 'preStressPreStressSteelClass',
  PreStressSteelDiameter = 'preStressPreStressSteelDiameter',
  PreStressAmount = 'preStressPreStressAmount',
  Date = 'preStressDate',
  Location = 'preStressLocation',
  User = 'preStressUser',
}